# src/integration-test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: api-licensed-business
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:49869
  description: Generated server url
paths:
  /{municipalityId}/restaurant-numbers:
    post:
      tags:
      - Restaurant Number Resources
      summary: "Create a restaurant number for an address, allocating the lowest free\
        \ number in the municipality"
      description: The restaurant number belongs to the premises and stays on the
        address when one business is replaced by the next.
      operationId: createRestaurantNumber
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: addressId
        in: query
        description: Address ID
        required: true
        schema:
          type: string
          minLength: 1
        example: 9ce333ec-a473-438b-8406-a71e957dc107
      responses:
        "201":
          description: Created
          headers:
            Location:
              style: simple
              schema:
                type: string
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/addresses:
    get:
      tags:
      - Address Resources
      summary: Get a paged list of addresses
      operationId: getAddresses
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: sortBy
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
            description: The properties to sort on
            examples:
            - propertyName
      - name: sortDirection
        in: query
        description: The sort order direction
        required: false
        schema:
          $ref: "#/components/schemas/Direction"
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          description: Page number
          examples:
          - 1
          minimum: 1
      - name: limit
        in: query
        description: Result size per page. Maximum allowed value is dynamically configured
        required: false
        schema:
          type: integer
          format: int32
          description: Result size per page. Maximum allowed value is dynamically
            configured
          examples:
          - 15
          minimum: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Addresses"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      tags:
      - Address Resources
      summary: Create an address
      operationId: createAddress
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Address"
        required: true
      responses:
        "201":
          description: Created
          headers:
            Location:
              style: simple
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/restaurant-numbers/{restaurantNumber}:
    get:
      tags:
      - Restaurant Number Resources
      summary: Get a restaurant number
      operationId: getRestaurantNumber
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: restaurantNumber
        in: path
        description: Restaurant number
        required: true
        schema:
          type: string
        example: 22813670
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RestaurantNumber"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/addresses/{addressId}:
    get:
      tags:
      - Address Resources
      summary: Get an address by ID
      operationId: getAddress
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: addressId
        in: path
        description: Address ID
        required: true
        schema:
          type: string
        example: 9ce333ec-a473-438b-8406-a71e957dc107
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Address"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/addresses/{addressId}/restaurant-numbers:
    get:
      tags:
      - Address Resources
      summary: Get every restaurant number at an address
      description: "Each restaurant number carries its status, ACTIVE when it has\
        \ an active assignment and otherwise AVAILABLE, and the premises name and\
        \ period of the assignment running today. A number without one shows its previous\
        \ assignment, and a number whose only assignments start later shows the next\
        \ one."
      operationId: getAddressRestaurantNumbers
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: addressId
        in: path
        description: Address ID
        required: true
        schema:
          type: string
        example: 9ce333ec-a473-438b-8406-a71e957dc107
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AddressRestaurantNumber"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/addresses/search:
    get:
      tags:
      - Address Resources
      summary: Search addresses
      operationId: searchAddresses
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: query
        in: query
        description: Search query
        required: true
        schema:
          type: string
          description: Search query
          examples:
          - Storgatan 1
          minLength: 1
      - name: sortBy
        in: query
        required: false
        schema:
          type: array
          items:
            type: string
            description: The properties to sort on
            examples:
            - propertyName
      - name: sortDirection
        in: query
        description: The sort order direction
        required: false
        schema:
          $ref: "#/components/schemas/Direction"
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          description: Page number
          examples:
          - 1
          minimum: 1
      - name: limit
        in: query
        description: Result size per page. Maximum allowed value is dynamically configured
        required: false
        schema:
          type: integer
          format: int32
          description: Result size per page. Maximum allowed value is dynamically
            configured
          examples:
          - 15
          minimum: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Addresses"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/addresses/lookup:
    get:
      tags:
      - Address Resources
      summary: Look up a single address by street address and postal code
      description: "Street address and postal code are matched on their normalized\
        \ form, so case, extra whitespace, a space in the postal code and a house\
        \ number letter written as \"1 A\" or \"1A\" all resolve to the same address.\
        \ This is the same key the duplicate check on create uses."
      operationId: lookupAddress
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: streetAddress
        in: query
        description: "Street address, written as in the address register"
        required: true
        schema:
          type: string
          description: "Street address, written as in the address register"
          examples:
          - Storgatan 1
          minLength: 1
      - name: postalCode
        in: query
        description: "Postal code, with or without space"
        required: true
        schema:
          type: string
          description: "Postal code, with or without space"
          examples:
          - 852 30
          minLength: 1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Address"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad Gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        detail:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
...
    Address:
      type: object
      description: Address model
      properties:
        id:
          type: string
          description: Address ID
          examples:
          - 9ce333ec-a473-438b-8406-a71e957dc107
          readOnly: true
        streetAddress:
          type: string
          description: Street address
          examples:
          - Storgatan 1
          minLength: 1
        postalCode:
          type: string
          description: Postal code
          examples:
          - 852 30
          minLength: 1
        postalArea:
          type: string
          description: Postal area
          examples:
          - Sundsvall
        municipalityId:
          type: string
          description: Municipality ID
          examples:
          - "2281"
          readOnly: true
        created:
          type: string
          format: date-time
          description: Timestamp when the address was created
          readOnly: true
      required:
      - postalCode
      - streetAddress
...
    RestaurantNumber:
      type: object
      description: Restaurant number model
      properties:
        id:
          type: string
          description: Restaurant number ID
          examples:
          - 9ce333ec-a473-438b-8406-a71e957dc107
        number:
          type: string
          description: Restaurant number
          examples:
          - "1001"
        municipalityId:
          type: string
          description: Municipality ID
          examples:
          - "2281"
        reported:
          type: boolean
          description: Whether the restaurant number has been reported to the Public
            Health Agency of Sweden
          examples:
          - false
        created:
          type: string
          format: date-time
          description: Timestamp when the restaurant number was created
    Direction:
      type: string
      description: The sort order direction
      enum:
      - ASC
      - DESC
      examples:
      - ASC
    Addresses:
      type: object
      description: Addresses model
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/Address"
            readOnly: true
        _meta:
          $ref: "#/components/schemas/PagingMetaData"
          readOnly: true
    PagingMetaData:
      type: object
      description: PagingMetaData model
      properties:
        page:
          type: integer
          format: int32
          description: Current page
          examples:
          - 5
          readOnly: true
        limit:
          type: integer
          format: int32
          description: Displayed objects per page
          examples:
          - 20
          readOnly: true
        count:
          type: integer
          format: int32
          description: Displayed objects on current page
          examples:
          - 13
          readOnly: true
        totalRecords:
          type: integer
          format: int64
          description: Total amount of hits based on provided search parameters
          examples:
          - 98
          readOnly: true
        totalPages:
          type: integer
          format: int32
          description: Total amount of pages based on provided search parameters
          examples:
          - 23
          readOnly: true
    AddressRestaurantNumber:
      type: object
      description: "Restaurant number at an address, with its status and the premises\
        \ name and period of its current assignment. A number without a current assignment\
        \ shows its previous one, and a number whose only assignments start later\
        \ shows the next one."
      properties:
        id:
          type: string
          description: Restaurant number ID
          examples:
          - 9ce333ec-a473-438b-8406-a71e957dc107
        number:
          type: string
          description: Restaurant number
          examples:
          - "22810001"
        municipalityId:
          type: string
          description: Municipality ID
          examples:
          - "2281"
        status:
          type: string
          description: "ACTIVE when the restaurant number has an active assignment,\
            \ otherwise AVAILABLE"
          enum:
          - ACTIVE
          - AVAILABLE
          examples:
          - ACTIVE
        premisesName:
          type: string
          description: "Name of the premises on the shown assignment, empty when the\
            \ number has never been assigned"
          examples:
          - Harrys Pub
        validFrom:
          type: string
          format: date
          description: First day of the shown assignment
          examples:
          - 2026-01-01
        validTo:
          type: string
          format: date
          description: "Last day of the shown assignment, empty when it is open ended"
          examples:
          - 2026-12-31
        created:
          type: string
          format: date-time
          description: Timestamp when the restaurant number was created
  securitySchemes: {}